jsobject editor extract context menu
