+++ app/client/src/pages/Editor/JSEditor/index.tsx
@@ -1,8 +1,7 @@
-import React from "react";
+import React, { useMemo } from "react";
 import type { RouteComponentProps } from "react-router";
 import type { JSCollection } from "entities/JSCollection";
-import type { AppState } from "@appsmith/reducers";
-import { connect } from "react-redux";
+import { useSelector } from "react-redux";
 import JsEditorForm from "./Form";
 import * as Sentry from "@sentry/react";
 import { getJSCollectionById } from "selectors/editorSelectors";
@@ -10,6 +9,7 @@ import CenteredWrapper from "components/designSystems/appsmith/CenteredWrapper";
 import Spinner from "components/editorComponents/Spinner";
 import styled from "styled-components";
 import EntityNotFoundPane from "../EntityNotFoundPane";
+import AppJSEditorContextMenu from "./AppJSEditorContextMenu";
 
 const LoadingContainer = styled(CenteredWrapper)`
   height: 50%;
@@ -22,31 +22,37 @@ interface ReduxStateProps {
 type Props = ReduxStateProps &
   RouteComponentProps<{ apiId: string; pageId: string }>;
 
-class JSEditor extends React.Component<Props> {
-  render() {
-    const { isCreating, jsCollection } = this.props;
-    if (isCreating) {
-      return (
-        <LoadingContainer>
-          <Spinner size={30} />
-        </LoadingContainer>
-      );
+function JSEditor(props: Props) {
+  const { pageId } = props.match.params;
+  const jsCollection = useSelector((state) =>
+    getJSCollectionById(state, props),
+  );
+  const { isCreating } = useSelector((state) => state.ui.jsPane);
+
+  const contextMenu = useMemo(() => {
+    if (!jsCollection) {
+      return null;
     }
-    if (!!jsCollection) {
-      return <JsEditorForm jsCollection={jsCollection} />;
-    }
-    return <EntityNotFoundPane />;
+
+    return (
+      <AppJSEditorContextMenu jsCollection={jsCollection} pageId={pageId} />
+    );
+  }, [jsCollection, pageId]);
+
+  if (isCreating) {
+    return (
+      <LoadingContainer>
+        <Spinner size={30} />
+      </LoadingContainer>
+    );
   }
+
+  if (!!jsCollection) {
+    return (
+      <JsEditorForm contextMenu={contextMenu} jsCollection={jsCollection} />
+    );
+  }
+  return <EntityNotFoundPane />;
 }
 
-const mapStateToProps = (state: AppState, props: Props): ReduxStateProps => {
-  const jsCollection = getJSCollectionById(state, props);
-  const { isCreating } = state.ui.jsPane;
-
-  return {
-    jsCollection,
-    isCreating: isCreating,
-  };
-};
-
-export default Sentry.withProfiler(connect(mapStateToProps)(JSEditor));
+export default Sentry.withProfiler(JSEditor);
